builds out a few modules
