+++ src/js/modules/settings.js
@@ -1,77 +1,3 @@
-(function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
-/*||
-||||    EARTHTAB CHROME EXTENSION
-||||
-||||    Scrapes /r/earthporn and makes the top image
-||||    your new tab background image.
-||||
-||||    v0.4.0 (Beta)
-||||
-||||    Dave Lunny 2015 (c) Licenced under MIT
-||||
-||||*/
-
-
-
-module.exports = (function(){
-
-  var $ = require('./modules/utils.js')();
-
-  var Clock = require('./modules/clock.js')('.js-time');
-      Clock.setClock();
-
-  var Settings = require('./modules/settings.js')();
-      Settings.init(function(){
-        console.log("Cool bro");
-      });
-
-      
-
-
-
-
-})();
-
-},{"./modules/clock.js":2,"./modules/settings.js":3,"./modules/utils.js":4}],2:[function(require,module,exports){
-/*||
-||||   Module::Clock
-||||
-||||   Responsible for setting/updating the clock
-||||   based on the element selector passed to the function
-*/
-
-
-var Clock = function( element ){
-  return {
-    el: document.querySelector(element),
-    /**   @name:    setClock
-      *   @params:  oldTime [number, time; optional]
-      *   @desc:    recursivly checks the time and alters it in the DOM
-      *             essentially a self-contained worker function
-      */
-    setClock: function( oldTime ){
-      var t = new Date(),
-          h = t.getHours(),
-          m = t.getMinutes();
-          if (m<10) { m = "0"+m; }
-          if( h > 12 ){
-            h = h - 12;
-          }
-      var time = h+":"+m;
-
-      if(!oldTime || oldTime !== time)
-        this.el.innerHTML = time;
-
-      var timeout = setTimeout(function(){
-        this.setClock(this.el, time);
-      }.bind(this), 5000);// five seconds is a lot but I'd rather that then taking the performance hit
-    }
-
-  }
-};
-
-module.exports = Clock;
-},{}],3:[function(require,module,exports){
 /*||
 ||||   Module::Settings
 ||||
@@ -319,62 +245,3 @@ var Settings = function(){
 };
 
 module.exports = Settings;
-},{"./utils.js":4}],4:[function(require,module,exports){
-/*||
-||||   Module::Utils
-||||
-||||   Some utility functions cause yay
-*/
-
-
-var Utils = function(){
-  return {
-    /**   @name:    resolveElement
-      *   @params:  el [string, selector OR DOM element object]
-      *   @desc:    handy utility that returns the actual element, whether passed a selector string or actual element
-      */
-    resolveElement: function( el ){
-      var element;
-      if( typeof el === 'string' )
-        element = document.querySelector(el);
-      else if( typeof el === 'object' )
-        //  TODO: check that it's a real DOM object
-        element = el;
-      else
-        throw "Not a real element!\nPlease pass either a selector string or element object!";
-      return element;
-    },
-    /**   @name:    addClass
-      *   @params:  el [string, selector], class [string]
-      *   @desc:    simplifies adding a class to an element
-      */
-    addClass: function( el, className ){
-      var element;
-      if( typeof el === 'object' )
-        element = el;
-      else if( typeof el === 'string' )
-        element = document.querySelector( el );
-      else
-        throw "addClass() needs either an element or a selector string!";
-      element.classList.add(className);
-    },
-    /**   @name:    removeClass
-      *   @params:  el [string, selector], class [string]
-      *   @desc:    simplifies removing a class from an element
-      */
-    removeClass: function( el, className ){
-      var element;
-      if( typeof el === 'object' )
-        element = el;
-      else if( typeof el === 'string' )
-        element = document.querySelector( el );
-      else
-        throw "removeClass() needs either an element or a selector string!";
-      element.classList.remove(className);
-    }
-  }
-};
-
-
-module.exports = Utils;
-},{}]},{},[1]);
