+++ src/js/modules/utils.js
@@ -1,0 +1,57 @@
+/*||
+||||   Module::Utils
+||||
+||||   Some utility functions cause yay
+*/
+
+
+var Utils = function(){
+  return {
+    /**   @name:    resolveElement
+      *   @params:  el [string, selector OR DOM element object]
+      *   @desc:    handy utility that returns the actual element, whether passed a selector string or actual element
+      */
+    resolveElement: function( el ){
+      var element;
+      if( typeof el === 'string' )
+        element = document.querySelector(el);
+      else if( typeof el === 'object' )
+        //  TODO: check that it's a real DOM object
+        element = el;
+      else
+        throw "Not a real element!\nPlease pass either a selector string or element object!";
+      return element;
+    },
+    /**   @name:    addClass
+      *   @params:  el [string, selector], class [string]
+      *   @desc:    simplifies adding a class to an element
+      */
+    addClass: function( el, className ){
+      var element;
+      if( typeof el === 'object' )
+        element = el;
+      else if( typeof el === 'string' )
+        element = document.querySelector( el );
+      else
+        throw "addClass() needs either an element or a selector string!";
+      element.classList.add(className);
+    },
+    /**   @name:    removeClass
+      *   @params:  el [string, selector], class [string]
+      *   @desc:    simplifies removing a class from an element
+      */
+    removeClass: function( el, className ){
+      var element;
+      if( typeof el === 'object' )
+        element = el;
+      else if( typeof el === 'string' )
+        element = document.querySelector( el );
+      else
+        throw "removeClass() needs either an element or a selector string!";
+      element.classList.remove(className);
+    }
+  }
+};
+
+
+module.exports = Utils;
